+++ src/InvokeException.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+/**
+ * @project hbcontainer
+ */
+
+namespace HbLib\Container;
+
+
+class InvokeException extends ContainerException
+{
+
+}