+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## (unreleased)
 
+## 1.1.0
+
+* Implement `Container::call()`
+* Added an base Exception class: `ContainerException`
+* Use Relflection API when invoking definition factories.
+
 ## 1.0.1
 
 * fix: Added missing class `UnresolvedContainerException`